test(db): retire agent-run model-key transition validator
Promote the canonical Agent Run model-key column, constraint, catalog, caller, and behavior invariants into permanent fresh and upgraded schema validation.

Refs #28878
