+++ turbo/packages/db/scripts/agent-compose-consolidation-preflight-manifest.ts
@@ -511,7 +511,6 @@ export const EXPECTED_REPOSITORY_DEPENDENCIES = {
     "#28080|checkpoint-configuration-independence-runtime-manifest|removal-owner:#26938-stage-8",
     "#28080|checkpoint-v7-protected-partition|removal-owner:#26938-stage-8",
     "#28304|usage-pack-pending-snapshot-dirty-upgrade|removal-owner:#28372",
-    "#28679|agent-run-built-in-model-key-expand-mirror-bridge|removal-owner:#28368",
   ],
 } as const satisfies RepositoryDependencyManifest;
